+++ p4/aggregate/aggregate_tna.p4
@@ -160,6 +160,7 @@ control TpsAggIngress(
             data_forward;
         }
         size = TABLE_SIZE;
+        default_action = data_forward(1);
     }
 
     action data_drop() {
@@ -294,20 +295,21 @@ control TpsAggIngress(
             add_switch_id_t.apply();
         } else {
             // Add IP & Protocol specific data to new INT data
-            if (data_inspection_t.apply().hit) {
-                if (hdr.udp_int.isValid()) {
-                    insert_udp_int_for_udp();
-                }
-                if (hdr.ipv4.isValid()) {
-                    data_inspect_packet_ipv4();
-                    if (hdr.tcp.isValid()) {
-                        insert_udp_int_for_tcp_ipv4();
+            if (! hdr.arp.isValid()) {
+                if (data_inspection_t.apply().hit) {
+                    if (hdr.udp_int.isValid()) {
+                        insert_udp_int_for_udp();
                     }
-                }
-                else if (hdr.ipv6.isValid()) {
-                    data_inspect_packet_ipv6();
-                    if (hdr.tcp.isValid()) {
-                        insert_udp_int_for_tcp_ipv6();
+                    if (hdr.ipv4.isValid()) {
+                        data_inspect_packet_ipv4();
+                        if (hdr.tcp.isValid()) {
+                            insert_udp_int_for_tcp_ipv4();
+                        }
+                    } else if (hdr.ipv6.isValid()) {
+                        data_inspect_packet_ipv6();
+                        if (hdr.tcp.isValid()) {
+                            insert_udp_int_for_tcp_ipv6();
+                        }
                     }
                 }
             }
@@ -315,27 +317,19 @@ control TpsAggIngress(
 
         // Basic forwarding and drop logic
         if (data_drop_t.apply().miss) {
-            if (hdr.arp.isValid() && hdr.arp.opcode == (bit<16>)0x1) {
-                ig_dprsr_md.digest_type = DIGEST_TYPE_ARP;
-                if (ig_intr_md.ingress_port != 0x1) {
-                    data_forward(1);
-                }
-            } else {
-                if (data_forward_t.apply().miss) {
-                /*
-                 * Send a digest if MAC address is unknown or if it is known
-                 * but attached to a different port as long as it does not come
-                 * in port 1, which should only be a switch. Nodes should be
-                 * plugged into the others.
-                 */
+            if (hdr.arp.isValid() && hdr.arp.opcode == (bit<16>)0x1
+                    && ig_intr_md.ingress_port == (bit<9>)0x1) {
+                // ARP Request - multicast out to all configured nodes
+                ig_tm_md.mcast_grp_a = (bit<16>)0x1;
+            } else if (data_forward_t.apply().miss) {
+                if (hdr.arp.isValid() && hdr.arp.opcode == (bit<16>)0x1
+                        && ig_intr_md.ingress_port != (bit<9>)0x1) {
+                    ig_dprsr_md.digest_type = DIGEST_TYPE_ARP;
+                } else {
                     if (ig_intr_md.ingress_port > 1) {
-                        // Send to NB switch at port 1
                         smac.apply();
                         if (src_miss == true || src_move != 0) {
                             ig_dprsr_md.digest_type = DIGEST_TYPE_FWD;
-                        }
-                        if (ig_intr_md.ingress_port != 0x1) {
-                            data_forward(1);
                         }
                     }
                 }
