+++ playbooks/scenarios/lab_trial/all-pkt-flood.yml
@@ -1,0 +1,69 @@
+# Copyright (c) 2020 Cable Television Laboratories, Inc.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# Simple scenario where packets are sent through 3 devices and only the last
+# one will be demonstrating dropped packets
+---
+# Discover Forwarding Paths
+- import_playbook: discover-paths.yml
+
+# Discover SB path
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: inet
+    rec_host: host1
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+# Data Inspection scenarios for UDP and IPv4
+- import_playbook: data-inspection.yml
+  vars:
+    di_log_prfx: di-ae-dd
+    scenario_send_protocol: UDP
+    scenario_send_ip_version: 4
+    di_send_packet_count: 150
+    di_min_rec_packet_count: 99
+    di_send_loops: 2
+    di_send_loop_delay: 10
+
+# Data Inspection scenarios for UDP and IPv6 with AE for testing dropping packets
+- import_playbook: data-inspection.yml
+  vars:
+    di_log_prfx: di-ae-dd
+    scenario_send_protocol: UDP
+    scenario_send_ip_version: 6
+    di_send_packet_count: 150
+    di_min_rec_packet_count: 99
+    di_send_loops: 2
+    di_send_loop_delay: 10
+
+# Data Inspection scenarios for TCP and IPv4 with AE for testing dropping packets
+- import_playbook: data-inspection.yml
+  vars:
+    di_log_prfx: di-ae-dd
+    scenario_send_protocol: TCP
+    scenario_send_ip_version: 4
+    di_send_packet_count: 150
+    di_min_rec_packet_count: 99
+    di_send_loops: 2
+    di_send_loop_delay: 10
+
+# Data Inspection scenarios for TCP and IPv6 with AE for testing dropping packets
+- import_playbook: data-inspection.yml
+  vars:
+    di_log_prfx: di-ae-dd
+    scenario_send_protocol: TCP
+    scenario_send_ip_version: 6
+    di_send_packet_count: 150
+    di_min_rec_packet_count: 99
+    di_send_loops: 2
+    di_send_loop_delay: 10
