+++ playbooks/scenarios/lab_trial/all.yml
@@ -13,119 +13,14 @@
 # Simple scenario where packets are sent through 3 devices and only the last
 # one will be demonstrating dropped packets
 ---
-# Data Forward scenarios for UDP and IPv4
-- import_playbook: data-forward.yml
-  vars:
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 4
+# Data Forward scenarios
+- import_playbook: all-data-forward.yml
 
-# Data Forward scenarios for UDP and IPv6
-- import_playbook: data-forward.yml
-  vars:
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 6
+# Data Inspection scenarios
+- import_playbook: all-data-inspection.yml
 
-# TODO/FIXME - Fix TCP headers are egressing from core-eth2 with incorrect port values
-# Data Forward scenarios for TCP and IPv4
-- import_playbook: data-forward.yml
-  vars:
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 4
+# Data Drop scenarios
+- import_playbook: all-data-drop.yml
 
-# Data Forward scenarios for TCP and IPv6
-- import_playbook: data-forward.yml
-  vars:
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 6
-
-# Data Inspection scenarios for UDP and IPv4
-- import_playbook: data-inspection.yml
-  vars:
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 4
-
-# Data Inspection scenarios for UDP and IPv6
-- import_playbook: data-inspection.yml
-  vars:
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 6
-
-# Data Inspection scenarios for TCP and IPv4
-- import_playbook: data-inspection.yml
-  vars:
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 4
-
-# Data Inspection scenarios for TCP and IPv6
-- import_playbook: data-inspection.yml
-  vars:
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 6
-
-# Data Drop scenarios for UDP and IPv4
-- import_playbook: data-drop.yml
-  vars:
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 4
-
-# Data Drop scenarios for UDP and IPv6
-- import_playbook: data-drop.yml
-  vars:
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 6
-
-# Data Drop scenarios for TCP and IPv4
-- import_playbook: data-drop.yml
-  vars:
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 4
-
-# Data Drop scenarios for TCP and IPv6
-- import_playbook: data-drop.yml
-  vars:
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 6
-
-# Data Inspection scenarios for UDP and IPv4
-- import_playbook: data-inspection.yml
-  vars:
-    di_log_prfx: di-ae-dd
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 4
-    di_send_packet_count: 150
-    di_min_rec_packet_count: 99
-    di_send_loops: 2
-    di_send_loop_delay: 10
-
-# Data Inspection scenarios for UDP and IPv6 with AE for testing dropping packets
-- import_playbook: data-inspection.yml
-  vars:
-    di_log_prfx: di-ae-dd
-    scenario_send_protocol: UDP
-    scenario_send_ip_version: 6
-    di_send_packet_count: 150
-    di_min_rec_packet_count: 99
-    di_send_loops: 2
-    di_send_loop_delay: 10
-
-# Data Inspection scenarios for TCP and IPv4 with AE for testing dropping packets
-- import_playbook: data-inspection.yml
-  vars:
-    di_log_prfx: di-ae-dd
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 4
-    di_send_packet_count: 150
-    di_min_rec_packet_count: 99
-    di_send_loops: 2
-    di_send_loop_delay: 10
-
-# Data Inspection scenarios for TCP and IPv6 with AE for testing dropping packets
-- import_playbook: data-inspection.yml
-  vars:
-    di_log_prfx: di-ae-dd
-    scenario_send_protocol: TCP
-    scenario_send_ip_version: 6
-    di_send_packet_count: 150
-    di_min_rec_packet_count: 99
-    di_send_loops: 2
-    di_send_loop_delay: 10
+# Packet Flood scenarios
+- import_playbook: all-pkt-flood.yml
