+++ playbooks/scenarios/lab_trial/discover-paths.yml
@@ -1,0 +1,80 @@
+# Copyright (c) 2020 Cable Television Laboratories, Inc.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# Simple scenario where packets are sent through 3 devices and only the last
+# one will be demonstrating dropped packets
+---
+# Discover NB paths
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: host1
+    rec_host: inet
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: host1
+    rec_host: ae
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: host2
+    rec_host: inet
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: host2
+    rec_host: ae
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+# Discover SB paths
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: inet
+    rec_host: host1
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: inet
+    rec_host: host2
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: ae
+    rec_host: host1
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- import_playbook: ../general/arping.yml
+  vars:
+    topo_dict: "{{ lookup('file','{{ topo_file_loc }}') | from_yaml }}"
+    send_host: ae
+    rec_host: host2
+    receiver: "{{ topo_dict.hosts[rec_host] }}"
+
+- hosts: localhost
+  tasks:
+    - name: Wait a few seconds
+      pause:
+        seconds: 5
+
+
